requirements.txt
iogram
pillow
google-generativeai
